Fetch navbar auth state on server
