+++ apps/web/src/app/global-error.tsx
@@ -6,7 +6,8 @@ import { Title, Text } from "@dotkomonline/ui"
 import PlausibleProvider from "next-plausible"
 import { QueryProvider } from "@/utils/trpc/QueryProvider"
 import { Footer } from "@/components/Footer/Footer"
-import { Navbar } from "@/components/Navbar/Navbar"
+import { NavbarContent } from "@/components/Navbar/NavbarContent"
+import { Auth0Provider } from "@auth0/nextjs-auth0/client"
 import { ThemeProvider } from "next-themes"
 import { setDefaultOptions as setDateFnsDefaultOptions } from "date-fns"
 import { nb } from "date-fns/locale"
@@ -27,27 +28,29 @@ export default function GlobalError({ error }: GlobalErrorProps) {
     <html lang="no">
       <body className="scrollbar-gutter-both overflow-x-clip bg-white dark:bg-stone-900">
         <PlausibleProvider domain="online.ntnu.no">
-          <QueryProvider>
-            <ThemeProvider defaultTheme="system" enableSystem attribute="data-theme">
-              <div className="page-shell">
-                <Navbar />
-                <main className="grow flex flex-col gap-8">
-                  <Title>En feil oppsto under innlasting av innhold.</Title>
+          <Auth0Provider>
+            <QueryProvider>
+              <ThemeProvider defaultTheme="system" enableSystem attribute="data-theme">
+                <div className="page-shell">
+                  <NavbarContent />
+                  <main className="grow flex flex-col gap-8">
+                    <Title>En feil oppsto under innlasting av innhold.</Title>
 
-                  <div className="rounded-lg p-4 bg-red-300 dark:bg-red-800">
-                    <Text>
-                      <strong>Detaljer:</strong> {error.message} {error.digest && `(digest=${error.digest})`}
+                    <div className="rounded-lg p-4 bg-red-300 dark:bg-red-800">
+                      <Text>
+                        <strong>Detaljer:</strong> {error.message} {error.digest && `(digest=${error.digest})`}
+                      </Text>
+                    </div>
+
+                    <Text className="text-gray-800 dark:text-stone-400" size="lg">
+                      Om feilen vedvarer, vennligst prøv å slette cookies eller refreshe siden.
                     </Text>
-                  </div>
-
-                  <Text className="text-gray-800 dark:text-stone-400" size="lg">
-                    Om feilen vedvarer, vennligst prøv å slette cookies eller refreshe siden.
-                  </Text>
-                </main>
-                <Footer />
-              </div>
-            </ThemeProvider>
-          </QueryProvider>
+                  </main>
+                  <Footer />
+                </div>
+              </ThemeProvider>
+            </QueryProvider>
+          </Auth0Provider>
         </PlausibleProvider>
       </body>
     </html>
